+++ app/controllers/api-v2-controller.js
@@ -5,6 +5,7 @@ var instanceModel = require( '../models/instance-model' );
 var account = require( '../models/account-model' );
 var auth = require( 'basic-auth' );
 var express = require( 'express' );
+var utils = require( '../lib/utils' );
 var router = express.Router();
 var quotaErrorMessage = 'Forbidden. No quota left';
 // var debug = require( 'debug' )( 'api-controller-v2' );
@@ -25,6 +26,7 @@ router
     .all( '/*/iframe', _setIframe )
     .all( '/survey/all', _setIframe )
     .all( '/surveys/list', _setIframe )
+    .all( '*/single*', _setMultipleAllowed )
     .all( '/survey/preview*', function( req, res, next ) {
         req.webformType = 'preview';
         next();
@@ -64,6 +66,10 @@ router
     .post( '/survey/offline', getNewOrExistingSurvey )
     .post( '/survey/iframe', getNewOrExistingSurvey )
     .delete( '/survey', deactivateSurvey )
+    .get( '/survey/single', getExistingSurvey )
+    .get( '/survey/single/iframe', getExistingSurvey )
+    .post( '/survey/single', getNewOrExistingSurvey )
+    .post( '/survey/single/iframe', getNewOrExistingSurvey )
     .get( '/survey/preview', getExistingSurvey )
     .get( '/survey/preview/iframe', getExistingSurvey )
     .post( '/survey/preview', getNewOrExistingSurvey )
@@ -313,6 +319,11 @@ function _setIframe( req, res, next ) {
     next();
 }
 
+function _setMultipleAllowed( req, res, next ) {
+    req.multipleAllowed = req.body.allow_multiple === 'true' || req.query.allow_multiple === 'true';
+    next();
+}
+
 function _setReturnQueryParam( req, res, next ) {
     var returnUrl = req.body.return_url || req.query.return_url;
 
@@ -344,6 +355,7 @@ function _generateWebformUrls( id, req ) {
     var baseUrl = protocol + '://' + req.headers.host + req.app.get( 'base path' ) + '/';
     var idPartOnline = '::' + id;
     var idPartOffline = '#' + id;
+    var idPartSingle = '::' + ( req.multipleAllowed ? id : utils.insecureAes192Encrypt( id, req.app.get( 'less secure encryption key' ) ) );
     var queryParts;
 
     req.webformType = req.webformType || 'default';
@@ -361,19 +373,23 @@ function _generateWebformUrls( id, req ) {
         case 'single':
             queryParts = [ req.defaultsQueryParam, req.returnQueryParam ];
             if ( iframePart ) {
-                queryParts.push( req.parentWindowOrigin );
+                queryParts.push( req.parentWindowOriginParam );
             }
             queryString = _generateQueryString( queryParts );
+            obj[ 'single' + ( iframePart ? '_iframe' : '' ) + '_url' ] = baseUrl +
+                'single/' + iframePart + idPartSingle + queryString;
             break;
         case 'all':
             // non-iframe views
             queryString = _generateQueryString( [ req.defaultsQueryParam ] );
             obj.url = baseUrl + idPartOnline + queryString;
+            obj.single_url = baseUrl + idPartSingle + queryString;
             obj.offline_url = baseUrl + OFFLINEPATH + idPartOffline;
             obj.preview_url = baseUrl + 'preview/' + idPartOnline + queryString;
             // iframe views
             queryString = _generateQueryString( [ req.defaultsQueryParam, req.parentWindowOriginParam ] );
             obj.iframe_url = baseUrl + IFRAMEPATH + idPartOnline + queryString;
+            obj.single_iframe_url = baseUrl + IFRAMEPATH + idPartSingle + queryString;
             obj.preview_iframe_url = baseUrl + 'preview/' + IFRAMEPATH + idPartOnline + queryString;
             // rest
             obj.enketo_id = id;